requirements.txt
# Python dependencies for the try-on app
fashn
python-dotenv
